src/requirements.txt
google-genai
pyyaml
dotenv